requirements.txt
fastapi>=0.109.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
pillow>=10.0.0
uvicorn[standard]>=0.27.0
